Crate delay

Structs

Delayed

Enums

DelayState 🔒

Wraps an iterator of type I by creating it on the first call to next(). Creates the wrapped iterator by calling F.

Traits

Delay

Lets you call fibs.delay() instead of delay(fibs), if such is your wont.

Functions

delay

Wraps an iterator created by new on the first call to next().